+++ TeamProject/Poker/Interfaces/IPokerDatabase.cs
@@ -14,7 +14,7 @@ namespace Poker.Interfaces
         IList<ICharacter> HumanPlayers { get; set; }
         IList<ICharacter> BotPlayers { get; set; }
         IList<ICard> TableCards { get; set; }
-        ulong Pot { get; set; }
+        int Pot { get; set; }
         Dictionary<string, bool>  Stages { get; set; }
         IList<ICharacter> CurrPlayers { get; set; }
